Fix Connection header management
By default, if the Upgrade header was specified, then the connection will be forwarded as upgrade to the Docker.
Although, if not, the Connection header was ( by mistake? ) set to "Close" which was affecting some Dockers which were expecting Keep Alive, for example.
This way we make every Docker happy by just forwarding the Connection header to them directly, if it's not Upgraded.